add to home icon - blue on white version
note: this causes a white background on the favicon in the browser tab as nuxt/pwa generates these icons from a single icon source. setting the background color in the manifest seems to make no difference and could be worth opening an issue with the nuxt-community/pwa-module folks.

lastly - unrelated but included in this commit: removed the COB Design system references, as those were for test only.

closes #22